refactor: apiの処理を別ファイルに切り出した
