+++ server/chat/src/api.rs
@@ -1,0 +1,1 @@
+pub mod user;
